Added Kanni Dog Images
Kanni is a sighthound very popular in the southern parts of Tamil Nadu. Kanni(கன்னி) means maiden in Tamil and these breeds were accompanying aristocrat brides to their new homes. Black and tan is the most common color you get to see. It needs a lot of space and could live on porridges made out of millets. It’s extremely responsive and attentive to the master’s commands while hunting.